+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopFileSystemIntegrationTest.java
@@ -47,6 +47,7 @@ import static com.google.common.truth.Truth.assertThat;
 import static com.google.common.truth.Truth.assertWithMessage;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Arrays.stream;
+import static org.apache.hadoop.fs.CommonConfigurationKeysPublic.HADOOP_SECURITY_CREDENTIAL_PROVIDER_PATH;
 import static org.junit.Assert.assertThrows;
 
 import com.google.api.client.http.HttpResponseException;
@@ -98,6 +99,8 @@ import org.junit.runners.JUnit4;
 /** Integration tests for GoogleHadoopFileSystem class. */
 @RunWith(JUnit4.class)
 public class GoogleHadoopFileSystemIntegrationTest extends GoogleHadoopFileSystemTestBase {
+
+  private static final String PUBLIC_BUCKET = "gs://gcp-public-data-landsat";
 
   @ClassRule
   public static NotInheritableExternalResource storageResource =
@@ -1793,14 +1796,12 @@ public class GoogleHadoopFileSystemIntegrationTest extends GoogleHadoopFileSyste
 
   @Test
   public void unauthenticatedAccessToPublicBuckets_fsGsProperties() throws Exception {
-    String publicBucket = "gs://gcp-public-data-landsat";
-
     Configuration config = new Configuration();
     config.setEnum("fs.gs.auth.type", AuthenticationType.UNAUTHENTICATED);
 
-    FileSystem fs = FileSystem.get(new URI(publicBucket), config);
+    FileSystem fs = FileSystem.get(new URI(PUBLIC_BUCKET), config);
 
-    FileStatus[] fileStatuses = fs.listStatus(new Path(publicBucket));
+    FileStatus[] fileStatuses = fs.listStatus(new Path(PUBLIC_BUCKET));
 
     assertThat(
             ((GoogleHadoopFileSystem) fs)
@@ -1820,15 +1821,24 @@ public class GoogleHadoopFileSystemIntegrationTest extends GoogleHadoopFileSyste
 
   @Test
   public void unauthenticatedAccessToPublicBuckets_googleCloudProperties() throws Exception {
-    String publicBucket = "gs://gcp-public-data-landsat";
-
     Configuration config = new Configuration();
     config.setEnum("google.cloud.auth.type", AuthenticationType.UNAUTHENTICATED);
 
-    FileSystem fs = FileSystem.get(new URI(publicBucket), config);
+    FileSystem fs = FileSystem.get(new URI(PUBLIC_BUCKET), config);
 
-    FileStatus[] fileStatuses = fs.listStatus(new Path(publicBucket));
+    FileStatus[] fileStatuses = fs.listStatus(new Path(PUBLIC_BUCKET));
 
     assertThat(fileStatuses).isNotEmpty();
+  }
+
+  @Test
+  public void testInitializeCompatibleWithHadoopCredentialProvider() throws Exception {
+    Configuration config = loadConfig();
+
+    // This does not need to refer to a real bucket/path for the test.
+    config.set(HADOOP_SECURITY_CREDENTIAL_PROVIDER_PATH, "jceks://gs@foobar/test.jceks");
+
+    FileSystem.get(new URI(PUBLIC_BUCKET), config);
+    // Initialization successful with no exception thrown.
   }
 }